Move class SettingsGroupPages one directory up (fits better there)
