+++ GitUI/SettingsDialog/GroupSettingsPage.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Windows.Forms;
 
-namespace GitUI.SettingsDialog.Pages
+namespace GitUI.SettingsDialog
 {
     /// <summary>
     /// Page to group other pages
